functions/streaming/requirements.txt
google-cloud-pubsub==1.4.1
google-cloud-bigquery==1.24.0
google-cloud-storage==1.26.0
google-cloud-core==1.3.0
pytz==2019.3
